requirements.txt
PyJWT
argon2-cffi
SQLAlchemy
fastapi
uvicorn
sqlalchemy
pymysql
passlib[bcrypt]
python-dotenv
pydantic
gunicorn
